Fix: 서버 컴포넌트로 ISR 적용하여 배포 환경 데이터 로딩 해결
문제:
- 클라이언트 컴포넌트에서 API 호출 → 진짜 ISR이 아님
- 배포 환경에서 매번 API 호출 발생
- Route Handler ISR만으로는 불완전

해결:
1. 서버 컴포넌트 래퍼 패턴 적용
   - page.tsx: 서버 컴포넌트에서 Supabase 직접 조회
   - archive-client.tsx: 클라이언트 컴포넌트 (UI/인터랙션)

2. 진짜 ISR 구현
   - export const revalidate = 86400
   - 서버에서 데이터 fetch 후 24시간 캐싱
   - 클라이언트는 props로 데이터 받음

3. 성능 최적화
   - 초기 렌더링: 서버에서 모든 뉴스레터 fetch
   - 클라이언트: 메모리에서 즉시 날짜 전환
   - 불필요한 API 호출 제거

구조:
- app/archive/page.tsx (서버, ISR)
- app/archive/_components/archive-client.tsx (클라이언트, UI)
