+++ app/archive/page.tsx
@@ -1,445 +1,110 @@
-'use client';
+import { supabase } from '@/lib/supabase';
+import ArchiveClient from './_components/archive-client';
+import type { DateString, NewsletterArchive, StockData } from './_types/archive.types';
 
-import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import { motion, AnimatePresence } from 'framer-motion';
-import { Calendar, ChevronDown, Loader2, TrendingUp } from 'lucide-react';
-import AnimatedBackground from '@/components/animated-background';
-import MiniCalendar from './_components/mini-calendar';
-import NewsletterCard from './_components/newsletter-card';
-import useNewsletterData from './_hooks/use-newsletter-data';
-import type { DateString } from './_types/archive.types';
-import { formatDisplayDate } from './_utils/date-formatting';
+// ISR: 24시간마다 재검증
+export const revalidate = 86400;
 
 /**
- * 뉴스레터 아카이브 페이지 (Matrix 테마 - S++ 엔터프라이즈급 99/100)
- *
- * 기능:
- * - 애니메이션 코드 빗방울이 있는 Matrix 블랙 배경
- * - Emerald-500 악센트 색상 (완벽한 브랜드 일관성)
- * - 스캔라인 CRT 효과
- * - WCAG AAA 색상 대비 (모든 곳에서 8.5:1 이상)
- * - 모바일 캘린더를 위한 포커스 트랩 관리
- * - 키보드 단축키 (월 네비게이션을 위한 Cmd+Left/Right)
- * - 성능 최적화 (캘린더 상태를 위한 useRef)
- * - 반응형 레이아웃 (데스크톱: 사이드바 + 콘텐츠, 모바일: 접기 가능)
- * - 데이터 표시기가 있는 캘린더 기반 날짜 선택
- * - Matrix 미학이 적용된 글래스모피즘 카드
- * - 스태거 효과가 적용된 부드러운 애니메이션 (60fps 보장)
- * - 로딩, 에러 및 빈 상태
- * - WCAG 2.1 AAA 접근성
- * - 완전한 키보드 네비게이션 지원
+ * 주식 JSON 데이터 파싱 및 검증
  */
-export default function ArchivePage() {
-  // 캘린더 상태 관리 (useRef 안티패턴 대신 적절한 useState 사용)
-  const [calendarState, setCalendarState] = useState({
-    year: new Date().getFullYear(),
-    month: new Date().getMonth(),
-  });
+function parseAndValidateStocks(jsonString: string): StockData[] | null {
+  const MAX_JSON_SIZE = 1024 * 1024; // 1MB
+  if (jsonString.length > MAX_JSON_SIZE) {
+    console.error('[Archive] JSON too large:', jsonString.length);
+    return null;
+  }
 
-  const [selectedDate, setSelectedDate] = useState<DateString | null>(null);
-  const [isCalendarOpen, setIsCalendarOpen] = useState(false);
-  const calendarButtonRef = useRef<HTMLButtonElement>(null);
+  try {
+    const stocks = JSON.parse(jsonString);
 
-  const { newsletter, isLoading, error, availableDates } = useNewsletterData(selectedDate);
-
-  // O(1) 조회를 위해 availableDates를 Set으로 변환
-  const availableDatesSet = useMemo(
-    () => new Set(availableDates),
-    [availableDates]
-  );
-
-  // 마운트 시 가장 최근 사용 가능한 날짜 자동 선택
-  useEffect(() => {
-    if (availableDates.length > 0 && !selectedDate) {
-      const mostRecent = availableDates[0];
-      setSelectedDate(mostRecent);
-
-      const date = new Date(mostRecent);
-      setCalendarState({
-        year: date.getFullYear(),
-        month: date.getMonth(),
-      });
-    }
-  }, [availableDates, selectedDate]);
-
-  const handlePrevMonth = useCallback(() => {
-    setCalendarState(prev => ({
-      year: prev.month === 0 ? prev.year - 1 : prev.year,
-      month: prev.month === 0 ? 11 : prev.month - 1,
-    }));
-  }, []);
-
-  const handleNextMonth = useCallback(() => {
-    setCalendarState(prev => ({
-      year: prev.month === 11 ? prev.year + 1 : prev.year,
-      month: prev.month === 11 ? 0 : prev.month + 1,
-    }));
-  }, []);
-
-  // 월 네비게이션을 위한 키보드 단축키
-  useEffect(() => {
-    function handleKeyDown(e: KeyboardEvent) {
-      // Cmd+Left: 이전 달
-      if (e.key === 'ArrowLeft' && (e.metaKey || e.ctrlKey)) {
-        e.preventDefault();
-        handlePrevMonth();
-      }
-      // Cmd+Right: 다음 달
-      if (e.key === 'ArrowRight' && (e.metaKey || e.ctrlKey)) {
-        e.preventDefault();
-        handleNextMonth();
-      }
-      // Escape: 모바일 캘린더 닫기
-      if (e.key === 'Escape' && isCalendarOpen) {
-        e.preventDefault();
-        setIsCalendarOpen(false);
-        calendarButtonRef.current?.focus();
-      }
+    if (!Array.isArray(stocks)) {
+      console.error('[Archive] Invalid data: not an array');
+      return null;
     }
 
-    window.addEventListener('keydown', handleKeyDown);
-    return () => window.removeEventListener('keydown', handleKeyDown);
-  }, [isCalendarOpen, handlePrevMonth, handleNextMonth]);
-
-  // 모바일 캘린더를 위한 포커스 트랩
-  useEffect(() => {
-    if (!isCalendarOpen) return;
-
-    const focusableElements = document.querySelectorAll(
-      '#mobile-calendar button, #mobile-calendar a, #mobile-calendar [tabindex]:not([tabindex="-1"])'
-    );
-
-    if (focusableElements.length === 0) return;
-
-    const firstElement = focusableElements[0] as HTMLElement;
-    const lastElement = focusableElements[focusableElements.length - 1] as HTMLElement;
-
-    function handleTabKey(e: KeyboardEvent) {
-      if (e.key !== 'Tab') return;
-
-      if (e.shiftKey) {
-        // Shift+Tab
-        if (document.activeElement === firstElement) {
-          e.preventDefault();
-          lastElement.focus();
-        }
-      } else {
-        // Tab
-        if (document.activeElement === lastElement) {
-          e.preventDefault();
-          firstElement.focus();
-        }
-      }
+    if (stocks.length === 0 || stocks.length > 100) {
+      console.error('[Archive] Invalid array size:', stocks.length);
+      return null;
     }
 
-    document.addEventListener('keydown', handleTabKey);
+    const isValid = stocks.every((stock) => {
+      return (
+        stock &&
+        typeof stock === 'object' &&
+        !('__proto__' in stock) &&
+        !('constructor' in stock) &&
+        typeof stock.ticker === 'string' &&
+        stock.ticker.length > 0 &&
+        typeof stock.name === 'string' &&
+        stock.name.length > 0 &&
+        typeof stock.close_price === 'number' &&
+        stock.close_price > 0 &&
+        typeof stock.rationale === 'string' &&
+        stock.rationale.length > 0 &&
+        stock.signals &&
+        typeof stock.signals === 'object' &&
+        typeof stock.signals.trend_score === 'number' &&
+        typeof stock.signals.momentum_score === 'number' &&
+        typeof stock.signals.volume_score === 'number' &&
+        typeof stock.signals.volatility_score === 'number' &&
+        typeof stock.signals.pattern_score === 'number' &&
+        typeof stock.signals.sentiment_score === 'number' &&
+        typeof stock.signals.overall_score === 'number'
+      );
+    });
 
-    // 캘린더가 열릴 때 첫 번째 요소에 포커스
-    setTimeout(() => firstElement.focus(), 100);
+    if (!isValid) {
+      console.error('[Archive] Invalid data: missing required fields');
+      return null;
+    }
 
-    return () => document.removeEventListener('keydown', handleTabKey);
-  }, [isCalendarOpen]);
+    return stocks as StockData[];
+  } catch (err) {
+    console.error('[Archive] JSON parse error:', err instanceof Error ? err.message : 'Unknown error');
+    return null;
+  }
+}
 
-  const handleDateSelect = useCallback((dateString: DateString) => {
-    setSelectedDate(dateString);
-    setIsCalendarOpen(false);
-  }, []);
+/**
+ * 서버 컴포넌트: 아카이브 페이지
+ * - Supabase에서 직접 데이터 fetch
+ * - ISR로 24시간 캐싱
+ * - 클라이언트 컴포넌트에 props로 전달
+ */
+export default async function ArchivePage() {
+  // Supabase에서 발송된 모든 뉴스레터 가져오기
+  const { data, error } = await supabase
+    .from('newsletter_content')
+    .select('*')
+    .eq('is_sent', true)
+    .order('newsletter_date', { ascending: false });
 
-  return (
-    <div className="min-h-screen bg-black text-white relative overflow-hidden">
-      {/* Matrix 애니메이션 배경 */}
-      <AnimatedBackground />
+  if (error) {
+    console.error('[Archive] Failed to fetch newsletters:', error);
+  }
 
-      {/* 스캔라인 효과 (CRT 미학) */}
-      <div className="fixed inset-0 pointer-events-none z-[1] opacity-[0.04]">
-        <div
-          className="absolute inset-0 bg-[linear-gradient(transparent_50%,rgba(16,185,129,0.04)_50%)] bg-[length:100%_4px] animate-[matrix-scan_8s_linear_infinite]"
-          aria-hidden="true"
-        />
-      </div>
+  // 날짜별로 뉴스레터 매핑
+  const newsletters: Record<string, NewsletterArchive> = {};
+  const availableDates: DateString[] = [];
 
-      {/* 메인 콘텐츠 */}
-      <main className="relative z-10 py-16 lg:py-20">
-        <div className="mx-auto max-w-7xl px-4 py-12 sm:px-6 lg:px-8 lg:py-16">
-          {/* 페이지 헤더 */}
-          <motion.header
-            initial={{ opacity: 0, y: -30 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8, ease: [0.19, 1, 0.22, 1] }}
-            className="mb-12 text-center lg:text-left"
-          >
-            <motion.div
-              initial={{ opacity: 0, scale: 0.95 }}
-              animate={{ opacity: 1, scale: 1 }}
-              transition={{ duration: 0.6, delay: 0.1 }}
-            >
-              <h1 className="mb-4 text-4xl sm:text-5xl lg:text-6xl font-extralight tracking-tight leading-[0.95]">
-                <span className="block text-emerald-500/90 mb-2">Newsletter</span>
-                <span className="block font-normal bg-clip-text text-transparent bg-gradient-to-r from-emerald-400 via-teal-400 to-emerald-400">
-                  Archive
-                </span>
-              </h1>
-            </motion.div>
-            <motion.p
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              transition={{ duration: 0.6, delay: 0.3 }}
-              className="text-lg text-white font-light leading-relaxed"
-            >
-              과거 발송된 AI 주식 분석 뉴스레터를 날짜별로 확인하세요
-            </motion.p>
-          </motion.header>
+  if (data) {
+    for (const row of data) {
+      const stocks = parseAndValidateStocks(row.gemini_analysis);
 
-          <div className="flex flex-col gap-8 lg:flex-row lg:items-start">
-            {/* 데스크톱 캘린더 사이드바 */}
-            <motion.aside
-              initial={{ opacity: 0, x: -40 }}
-              animate={{ opacity: 1, x: 0 }}
-              transition={{ duration: 0.8, delay: 0.2, ease: [0.19, 1, 0.22, 1] }}
-              className="hidden lg:block lg:w-[340px] lg:flex-shrink-0"
-            >
-              <div className="sticky top-8 space-y-6">
-                <MiniCalendar
-                  year={calendarState.year}
-                  month={calendarState.month}
-                  selectedDate={selectedDate}
-                  availableDates={availableDatesSet}
-                  onDateSelect={handleDateSelect}
-                  onPrevMonth={handlePrevMonth}
-                  onNextMonth={handleNextMonth}
-                />
+      if (stocks) {
+        const newsletter: NewsletterArchive = {
+          date: row.newsletter_date,
+          stocks,
+          sentAt: row.sent_at,
+          subscriberCount: row.subscriber_count ?? 0,
+        };
 
-                {/* 뉴스레터 헤더 - 캘린더 밑 */}
-                {!isLoading && newsletter && (
-                  <motion.header
-                    initial={{ opacity: 0, y: 20 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    transition={{ duration: 0.5, delay: 0.1 }}
-                    className="
-                      rounded-xl border border-emerald-500/20
-                      bg-slate-900/60 backdrop-blur-xl
-                      p-6 shadow-lg
-                    "
-                  >
-                    <div className="flex items-start justify-between">
-                      <div>
-                        <h2 className="mb-2 text-2xl font-bold text-white">
-                          {formatDisplayDate(newsletter.date)}
-                        </h2>
-                        {newsletter.sentAt && (
-                          <p className="text-sm text-slate-300 font-mono">
-                            발송: {new Date(newsletter.sentAt).toLocaleString('ko-KR')}
-                            {/*{newsletter.subscriberCount > 0 && (*/}
-                            {/*  <> • 구독자 {newsletter.subscriberCount.toLocaleString()}명</>*/}
-                            {/*)}*/}
-                          </p>
-                        )}
-                      </div>
-                      <TrendingUp
-                        className="h-8 w-8 text-emerald-400"
-                        aria-hidden="true"
-                      />
-                    </div>
-                  </motion.header>
-                )}
-              </div>
-            </motion.aside>
+        newsletters[row.newsletter_date] = newsletter;
+        availableDates.push(row.newsletter_date as DateString);
+      }
+    }
+  }
 
-            {/* 모바일 캘린더 토글 */}
-            <motion.div
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.6, delay: 0.3 }}
-              className="lg:hidden"
-            >
-              <button
-                ref={calendarButtonRef}
-                onClick={() => setIsCalendarOpen(!isCalendarOpen)}
-                aria-expanded={isCalendarOpen}
-                aria-controls="mobile-calendar"
-                aria-label={`캘린더 ${isCalendarOpen ? '닫기' : '열기'}`}
-                className="
-                  flex w-full items-center justify-between
-                  rounded-xl border border-emerald-500/20
-                  bg-slate-900/60 backdrop-blur-xl
-                  p-4 shadow-lg
-                  transition-all duration-300
-                  hover:border-emerald-500/40
-                  hover:bg-slate-900/80
-                  focus-visible:outline-none focus-visible:ring-2
-                  focus-visible:ring-emerald-500 focus-visible:ring-offset-2
-                  focus-visible:ring-offset-black
-                "
-              >
-                <div className="flex items-center gap-3">
-                  <Calendar className="h-5 w-5 text-emerald-400" aria-hidden="true" />
-                  <span className="font-semibold text-white">
-                    {formatDisplayDate(selectedDate)}
-                  </span>
-                </div>
-                <motion.div
-                  animate={{ rotate: isCalendarOpen ? 180 : 0 }}
-                  transition={{ duration: 0.3, ease: [0.19, 1, 0.22, 1] }}
-                >
-                  <ChevronDown className="h-5 w-5 text-slate-400" aria-hidden="true" />
-                </motion.div>
-              </button>
-
-              <AnimatePresence>
-                {isCalendarOpen && (
-                  <motion.div
-                    id="mobile-calendar"
-                    role="dialog"
-                    aria-modal="true"
-                    aria-label="캘린더"
-                    initial={{ opacity: 0, height: 0 }}
-                    animate={{ opacity: 1, height: 'auto' }}
-                    exit={{ opacity: 0, height: 0 }}
-                    transition={{ duration: 0.3, ease: [0.19, 1, 0.22, 1] }}
-                    className="mt-4 overflow-hidden"
-                  >
-                    <MiniCalendar
-                      year={calendarState.year}
-                      month={calendarState.month}
-                      selectedDate={selectedDate}
-                      availableDates={availableDatesSet}
-                      onDateSelect={handleDateSelect}
-                      onPrevMonth={handlePrevMonth}
-                      onNextMonth={handleNextMonth}
-                    />
-                  </motion.div>
-                )}
-              </AnimatePresence>
-            </motion.div>
-
-            {/* 콘텐츠 영역 */}
-            <motion.section
-              initial={{ opacity: 0, x: 40 }}
-              animate={{ opacity: 1, x: 0 }}
-              transition={{ duration: 0.8, delay: 0.4, ease: [0.19, 1, 0.22, 1] }}
-              className="flex-1 min-w-0"
-              aria-live="polite"
-              aria-busy={isLoading}
-            >
-              <AnimatePresence mode="sync">
-                {/* 로딩 상태 */}
-                {isLoading && (
-                  <motion.div
-                    key="loading"
-                    initial={{ opacity: 0, scale: 0.95 }}
-                    animate={{ opacity: 1, scale: 1 }}
-                    exit={{ opacity: 0, scale: 0.95 }}
-                    transition={{ duration: 0.3 }}
-                    className="flex min-h-[500px] items-center justify-center"
-                  >
-                    <div className="text-center">
-                      <Loader2
-                        className="mx-auto mb-6 h-12 w-12 animate-spin text-emerald-500"
-                        aria-hidden="true"
-                      />
-                      <p className="text-lg text-white font-light">
-                        뉴스레터를 불러오는 중...
-                      </p>
-                      <p className="text-sm text-slate-300 mt-2">
-                        약 2-3초 소요됩니다
-                      </p>
-                    </div>
-                  </motion.div>
-                )}
-
-                {/* Error State */}
-                {!isLoading && error && (
-                  <motion.div
-                    key="error"
-                    initial={{ opacity: 0, scale: 0.95 }}
-                    animate={{ opacity: 1, scale: 1 }}
-                    exit={{ opacity: 0, scale: 0.95 }}
-                    transition={{ duration: 0.3 }}
-                    role="alert"
-                    className="
-                      rounded-2xl border border-red-500/30
-                      bg-red-500/10 backdrop-blur-xl
-                      p-8 text-center
-                      shadow-[0_0_30px_rgba(239,68,68,0.2)]
-                    "
-                  >
-                    <p className="text-lg text-red-400 font-medium mb-4">{error}</p>
-                    <button
-                      onClick={() => window.location.reload()}
-                      className="px-4 py-2 rounded-lg bg-red-500/20 hover:bg-red-500/30 text-red-300 font-medium transition-colors"
-                    >
-                      다시 시도
-                    </button>
-                  </motion.div>
-                )}
-
-                {/* 빈 상태 */}
-                {!isLoading && !error && !newsletter && (
-                  <motion.div
-                    key="empty"
-                    initial={{ opacity: 0, scale: 0.95 }}
-                    animate={{ opacity: 1, scale: 1 }}
-                    exit={{ opacity: 0, scale: 0.95 }}
-                    transition={{ duration: 0.3 }}
-                    className="
-                      rounded-2xl border border-emerald-500/20
-                      bg-slate-900/60 backdrop-blur-xl
-                      p-12 text-center
-                      shadow-2xl
-                    "
-                  >
-                    <Calendar
-                      className="mx-auto mb-6 h-12 w-12 sm:h-16 sm:w-16 text-slate-600"
-                      aria-hidden="true"
-                    />
-                    <h2 className="mb-3 text-2xl font-bold text-white">
-                      날짜를 선택해주세요
-                    </h2>
-                    <p className="text-slate-300 leading-relaxed">
-                      {availableDates.length > 0
-                        ? '캘린더에서 날짜를 선택하면 해당 날짜의 뉴스레터를 확인할 수 있습니다'
-                        : '아직 발송된 뉴스레터가 없습니다'}
-                    </p>
-                  </motion.div>
-                )}
-
-                {/* 뉴스레터 콘텐츠 - 주식 카드 그리드 */}
-                {!isLoading && newsletter && (() => {
-                  // 모든 카드의 rationale 아이템 개수 중 최대값 계산
-                  const maxRationaleItems = Math.max(
-                    ...newsletter.stocks.map((stock) => stock.rationale.split('|').length)
-                  );
-
-                  // Overall Score 높은 순서대로 정렬
-                  const sortedStocks = [...newsletter.stocks].sort(
-                    (a, b) => b.signals.overall_score - a.signals.overall_score
-                  );
-
-                  return (
-                    <motion.div
-                      key={newsletter.date}
-                      initial={{ opacity: 0, y: 30 }}
-                      animate={{ opacity: 1, y: 0 }}
-                      exit={{ opacity: 0, y: -30 }}
-                      transition={{ duration: 0.5, ease: [0.19, 1, 0.22, 1] }}
-                      className="grid gap-8 md:grid-cols-2 xl:grid-cols-3"
-                    >
-                      {sortedStocks.map((stock, index) => (
-                        <NewsletterCard
-                          key={stock.ticker}
-                          stock={stock}
-                          index={index}
-                          maxRationaleItems={maxRationaleItems}
-                        />
-                      ))}
-                    </motion.div>
-                  );
-                })()}
-              </AnimatePresence>
-            </motion.section>
-          </div>
-        </div>
-      </main>
-    </div>
-  );
+  return <ArchiveClient initialNewsletters={newsletters} availableDates={availableDates} />;
 }
